Basic API Connection Done

+++ lib/states/models/driver_details/driver_details.dart
@@ -45,7 +45,7 @@ class DriverDetailsModel {
   int cityId;
   String username;
   String image;
-  int commission;
+  double commission;
 
   factory DriverDetailsModel.fromJson(Map<String, dynamic> json) =>
       DriverDetailsModel(
